requirements.txt
openai>=1.0.0
langchain>=0.1.0,<0.2.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
langchain-text-splitters>=0.0.1
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
cohere>=4.0.0
streamlit>=1.32.0
pypdf>=3.0.0
python-docx>=1.1.0
tiktoken>=0.6.0
python-dotenv>=1.0.0
